Fix navigation blocked notification - i18n and consistent UI
## Fixes
- Added proper i18n support for navigation blocked notifications
- Both English and zh_TW (Traditional Chinese) translations included
- Extracted and reused existing notification system from options page

## Changes
- Created shared notification-utils.js for consistent notifications across extension
- Added i18n keys: 'navigationBlocked' and 'completeTheMazeToVisit'
- Updated maze.js to use shared showNotification with proper localization
- Removed duplicate notification code from options.js
- Used 'warning' type for navigation blocked notifications (orange styling)

## Benefits
- Consistent notification styling across all extension pages
- Proper localization support (zh_TW: '導航已被阻止', '完成迷宮以造訪 {site}')
- Reduced code duplication and maintenance burden
- Better UX with consistent notification behavior

Addresses feedback on PR #5
